+++ code/weapons/StoneHatchet.cs
@@ -14,7 +14,7 @@ partial class StoneHatchet : MeleeWeapon
 	public override string ViewModelPath => "models/tools/hatchet/hatchet_v.vmdl";
 	public override float PrimaryRate => 1.0f;
 
-	public override Dictionary<ResourceType, int> ResourceYield => new()
+	public override Dictionary<ResourceType, float> ResourceYield => new()
 	{
 		{ ResourceType.Wood, 1 },
 		{ ResourceType.Stone, 0 }
